chore: refactor create email validator
