+++ Ivy/Views/Forms/FormHelpers.cs
@@ -104,34 +104,6 @@ public static class FormHelpers
         var nullabilityInfo = nullabilityContext.Create(fieldInfo);
         return nullabilityInfo.ReadState != NullabilityState.Nullable;
     }
-
-    /// <summary> Creates an email validator using EmailAddressAttribute for proper email validation. </summary>
-    public static Func<object?, (bool, string)> CreateEmailValidator(string fieldName)
-    {
-        var emailValidator = new EmailAddressAttribute();
-        return email =>
-        {
-            if (email is not string emailStr || string.IsNullOrWhiteSpace(emailStr))
-                return (true, ""); // Empty is handled by Required validator
-
-            try
-            {
-                var validationContext = new ValidationContext(new { })
-                {
-                    MemberName = fieldName,
-                    DisplayName = fieldName
-                };
-                var result = emailValidator.GetValidationResult(emailStr, validationContext);
-                return result == ValidationResult.Success
-                    ? (true, "")
-                    : (false, result?.ErrorMessage ?? "Please enter a valid email address");
-            }
-            catch
-            {
-                return (false, "Please enter a valid email address");
-            }
-        };
-    }
 }
